GC pressure

Computes the following equation: $$\frac{\text{W}\to\text{S} \quad-\quad \text{S}\to\text{W}}{\text{W}\to\text{S} \quad+\quad \text{S}\to\text{W}}$$

Usage

GCpressure(m)

Arguments

m

A matrix of counts or probabilities for bases of the target genome to be aligned to bases on the query genome. As a convenience it can also receive a list produced by the readTrainFile() function, containing this matrix.

Value

A numeric value between \(-1\) and \(1\).

Note

This index was suggested to by by M365 Copilot while using it to write the GCequilibrium function and its documentation. Despite my attempts I could not find a paper that uses it explicitely… I am including it in this package because I am screening as many indices that I can extract from an alignment pair.

See also

Author

Charles Plessy

Examples

GCpressure(exampleSubstitutionMatrix)
#> [1] 0.02888339
GCpressure(t(exampleSubstitutionMatrix))
#> [1] -0.02888339