Gene Order Conservation

The Gene Order Conservation (GOC) number defined by Rocha (2003) is: “the average number of orthologues for which the consecutive orthologue co-occurs close by in the other genome. It varies between 0 (no co-occurrence) and 1 (complete gene order conservation)”.

Usage

GOC(gb, vicinity = 4, debug = FALSE)

Arguments

gb

A GBreaks object.

vicinity

How far to search for the neighbour orthologue.

debug

See below.

Value

Returns a numeric value between 0 and 1. If debug = TRUE, returns a copy of the gb object with additional columns showing details of the computation.

Note

Note that calculating GOC on whole-genome alignments is not expected to produce meaningful results. This function is more useful when comparing the position of orthologues, represented in a GBreaks object.

References

Rocha, Eduardo P C. “DNA repeats lead to the accelerated loss of gene order in bacteria.” Trends in genetics : TIG vol. 19,11 (2003): 600-3. doi:10.1016/j.tig.2003.09.011

See also

Examples

exampleColinear
#> GBreaks object with 2 ranges and 1 metadata column:
#>       seqnames    ranges strand |        query
#>          <Rle> <IRanges>  <Rle> |    <GRanges>
#>   [1]     chrA   100-150      + | chrB:100-150
#>   [2]     chrA   251-300      + | chrB:251-300
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome
GOC(exampleColinear)
#> [1] 1

exampleTranslocation
#> GBreaks object with 3 ranges and 1 metadata column:
#>       seqnames    ranges strand |        query
#>          <Rle> <IRanges>  <Rle> |    <GRanges>
#>   [1]     chrA   100-200      + | chrB:100-200
#>   [2]     chrA   201-300      + | chrC:201-300
#>   [3]     chrA   301-400      + | chrB:301-400
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome
GOC(exampleTranslocation)
#> [1] 0.5
GOC(exampleTranslocation, v=1)
#> [1] 0

# GOC computation is strandless
exampleInversion
#> GBreaks object with 3 ranges and 1 metadata column:
#>       seqnames    ranges strand |        query
#>          <Rle> <IRanges>  <Rle> |    <GRanges>
#>   [1]     chrA   100-190      + | chrB:100-190
#>   [2]     chrA   210-291      - | chrB:210-291
#>   [3]     chrA   301-400      + | chrB:301-400
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome
GOC(exampleInversion)
#> [1] 1