Percent difference Distance

The P-distance is simply the percentage of the aligned base pairs where the two bases are different. See May 2004 for a discussion on the many ways to compute percent identity (and therefore difference.)

Usage

P_distance(train_parameters)

Arguments

train_parameters

A list containing the probabilities of the alignment, produced by the readTrainFile() function.

Value

A numeric value representing the evolutionary distance between two genomes. The larger the value, the more divergent the genomes.

References

May, A. C. W. (2004). Percent Sequence Identity: The Need to Be Explicit. Structure, 12(5), 737–738. doi:doi.org/10.1016/j.str.2004.04.001

See also

Author

Charles Plessy

Examples

parameters <- readTrainFile(system.file("extdata/example.train", package = "GenomicBreaks"))
P_distance(parameters)
#> [1] 0.015976