Conversion from CNEr::Axt objects

Conversion from CNEr::Axt objects

Note

By default, in the Axt objects produced by the CNEr::readAxt function, the coordinates of the query genome are represented with the same numerical value as in the original file. In Axt files, when the alignment is on the minus strand, the query genome coordinates have their origin on the reverse-complement strand. Therefore, although in the CNEr object they are stored in a GRanges object, they do not represent genomic ranges. In order to do so, there is a CNEr::fixCoordinates function.

See also