Chain contigs in pairwise alignments

Description TBW

chain_contigs(gb, tol = Inf, drop.within = FALSE)

Arguments

gb
GBreaks object of the pairwise alignment.

tol
Unaligned region of width lesser than or equal to tol in both the reference and query case will be bridged in coalescing.

drop.within
Discard the chains that are included in other chains.

Value

TBW

See also

Author

Charles Plessy

Examples

gb4       <- GRanges(c("Ref:100-200:+", "Ref:400-500:+", "Ref:600-700:+"))
gb4$query <- GRanges(c("Que:1100-1200:+", "Que:1700-1800:+", "Que:1500-1600:+"))
chain_contigs(gb4)
#> GBreaks object with 3 ranges and 2 metadata columns:
#>       seqnames    ranges strand |         query     score
#>          <Rle> <IRanges>  <Rle> |     <GRanges> <integer>
#>   [1]      Ref   100-200      + | Que:1100-1200       101
#>   [2]      Ref   400-500      + | Que:1700-1800       101
#>   [3]      Ref   600-700      + | Que:1500-1600       101
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths