Coalesce Pairwise Alignments

This algorithm take in a genome-to-genome alignment, represented as a collection of intervals in a query genome paired with intervals in a target genome. It reduces the number of pairs by coalescing pairs that are within close proximity on the same strand (user determined).

coalesce_contigs(gb, tol = Inf, minwidth = 0)

Arguments

gb
GBreaks object of the pairwise alignment.

tol
Unaligned region of width lesser than or equal to tol in both the reference and query case will be bridged in coalescing.

minwidth
Remove the intervals whose width smaller than this value.

Value

Returns a new GBreaks object with a reduced number of alignments fragments due to coalescion. The returned object is sorted ignoring strand. For convenience during analysis sessions, its score is set to the width of the ranges on the target genome.

Note

Fragmented alignments arising from incorrect basecalls, misassembly or misalignments can cause us to infer artificial breakpoints

Internally, the coalesce_contigs function uses the precede and follow functions of the GenomicRanges package. For a given range, these functions return the index position of the range it precedes or follows, or NA as the first range follows nothing and the last range precedes nothing. See the examples for details.

See also

Author

Charlotte West

Charles Plessy

Examples

# Ranges on the plus strand that should coalesce
gb1       <- GRanges(c(A="Ref:100-200:+", B="Ref:400-500:+"))
gb1$query <- GRanges(c(A="Que:100-200",   B="Que:400-500"))
gb1

#> GRanges object with 2 ranges and 1 metadata column:
#>     seqnames    ranges strand |       query
#>        <Rle> <IRanges>  <Rle> |   <GRanges>
#>   A      Ref   100-200      + | Que:100-200
#>   B      Ref   400-500      + | Que:400-500
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

coalesce_contigs(gb1)

#> GRanges object with 1 range and 2 metadata columns:
#>       seqnames    ranges strand |       query     score
#>          <Rle> <IRanges>  <Rle> |   <GRanges> <integer>
#>   [1]      Ref   100-500      + | Que:100-500       401
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Reference range [1] precedes reference range [2]
precede(gb1)

#> [1]  2 NA

# Query range [1] precedes query range [2]
precede(gb1$query)

#> [1]  2 NA

# Ranges on the minus strand that should coalesce
gb2       <- GRanges(c(B="Ref:100-200:-", A="Ref:400-500:-"))
gb2$query <- GRanges(c(B="Que:400-500",   A="Que:100-200"))
gb2$qname <- names(gb2$query)
gb2

#> GRanges object with 2 ranges and 2 metadata columns:
#>     seqnames    ranges strand |       query       qname
#>        <Rle> <IRanges>  <Rle> |   <GRanges> <character>
#>   B      Ref   100-200      - | Que:400-500           B
#>   A      Ref   400-500      - | Que:100-200           A
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Reference range [1] follows reference range [2]
follow(gb2)

#> [1]  2 NA

# Or, ignoring strand, reference range [1] precedes reference range [2]
precede(gb2, ignore.strand = TRUE)

#> [1]  2 NA

# Query range [1] follows query range [2]
follow(gb2$query)

#> [1]  2 NA

coalesce_contigs(gb2)

#> GRanges object with 1 range and 2 metadata columns:
#>       seqnames    ranges strand |       query     score
#>          <Rle> <IRanges>  <Rle> |   <GRanges> <integer>
#>   [1]      Ref   100-500      - | Que:100-500       401
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Ranges on the minus strand that should not coalesce because they are not
# ordered properly
gb3       <- GRanges(c("Ref:100-200:-", "Ref:400-500:-"))
gb3$query <- GRanges(c("Que:100-200",   "Que:400-500"))
# Reference range [1] follows reference range [2]
follow(gb3)

#> [1]  2 NA

# Query range [1] follows query range [2]
precede(gb3$query)

#> [1]  2 NA

coalesce_contigs(gb3)

#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |       query     score
#>          <Rle> <IRanges>  <Rle> |   <GRanges> <integer>
#>   [1]      Ref   100-200      - | Que:100-200       101
#>   [2]      Ref   400-500      - | Que:400-500       101
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Ranges on the plus strand that should not coalesce because they are not
# ordered properly
gb4       <- GRanges(c("Ref:100-200:+", "Ref:400-500:+", "Ref:600-700:+"))
gb4$query <- GRanges(c("Que:1100-1200:+", "Que:1700-1800:+", "Que:1500-1600:+"))
coalesce_contigs(gb4)

#> GRanges object with 3 ranges and 2 metadata columns:
#>       seqnames    ranges strand |         query     score
#>          <Rle> <IRanges>  <Rle> |     <GRanges> <integer>
#>   [1]      Ref   100-200      + | Que:1100-1200       101
#>   [2]      Ref   400-500      + | Que:1700-1800       101
#>   [3]      Ref   600-700      + | Que:1500-1600       101
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths