Index representing how correlated syntenic regions are

A sequence feature on the target genome is considered ‘syntenic’ to the feature of the query genome matched by the most of its alignments. The correlation coefficient of the center of syntenic alignments is then calculated. Then, the computed numbers for all the target's features are averaged as a mean weighted by is the total number of aligned bases of each feature. Thus, a number close to 1 is expected for closely related genomes.

Usage

correlation_index(gb)

Arguments

gb

A GBreaks object.

Value

Returns a numeric value between 0 and 1.

Note

Here, the definition of ‘synteny’ is sequences on the same DNA strands, or in technical terms: sequences that are on the same feature, regardless of strand orientation.

See also

Examples

gb       <- GRanges(c("Ref:100-200:+",   "Ref:400-500:+",    "Ref:600-700:+"))
gb$query <- GRanges(c("Que:1100-1200:+", "Que2:1700-1800:+", "Que:1500-1600:+"))
correlation_index(gb)
#> [1] 1