Filter colinear regions

Filter colinear regions

filterColinearRegions(gb, rename = TRUE)

Arguments

gb
A GBreaks object processed with flagColinearAlignments.

rename
Replace range names by their numeric order before subsetting.

Value

Returns the GBreaks object in which all ranges that are not the central part of an inversion triplet have been discarded. If the object was missing the colinear metadata column, return the object after discarding all of its ranges.

See also

Examples

filterColinearRegions(flagColinearAlignments(exampleColinear))

#> GBreaks object with 2 ranges and 2 metadata columns:
#>     seqnames    ranges strand |        query  colinear
#>        <Rle> <IRanges>  <Rle> |    <GRanges> <logical>
#>   1     chrA   100-150      + | chrB:100-150      TRUE
#>   2     chrA   251-300      + | chrB:251-300     FALSE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome