Flag insersions

Flag ranges that start a triplet where the reference range is colinear and the query range is made of two blocks that would be colinear if they were not separated by an block on a different molecule.

flagInsersions(gr_ob, tol = Inf)

Arguments

gr_ob
A GenomicBreaks object.

tol
Tolerance window for the distance between two ranges.

Value

Returns the GenomicBreaks object with an extra ins metadata column.

Details

Here is a trivial example of an insersion.

┌───────────────┬───────────────┬───────────────┐
│ XSR:101-200:+ │ XSR:201-300:+ │ XSR:301-400:+ │  (OKI2018_I69)
└───────────────┴───────────────┴───────────────┘
┌───────────────┬───────────────┬───────────────┐
│  S1:101-200:+ │  S2:801-900:+ │  S1:201-300:+ │  (OSKA2016)
└───────────────┴───────────────┴───────────────┘

See also

Examples

ins       <- GRanges(c("XSR:101-200:+", "XSR:201-300:+",  "XSR:301-400:+"))
ins$query <- GRanges(c( "S1:101-200",   "S2:801-900",     "S1:201-300"   ))
flagInsersions(ins)

#> GRanges object with 3 ranges and 2 metadata columns:
#>       seqnames    ranges strand |      query       ins
#>          <Rle> <IRanges>  <Rle> |  <GRanges> <logical>
#>   [1]      XSR   101-200      + | S1:101-200      TRUE
#>   [2]      XSR   201-300      + | S2:801-900     FALSE
#>   [3]      XSR   301-400      + | S1:201-300     FALSE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths