Tau index

Tissue-specificity index from Yanai and coll., 2005 repurposed in 2022 by Ranz and coll. for synteny analysis.

Usage

tau_index(gb)

Arguments

gb

A GBreaks object.

Value

Returns a numeric value between 0 and 1.

Details

In 2022, Ranz and coll., the tau index was computed for each chromosome in a target genome compared to a query, with a formula such as: for a given target feature, count its one-to-one orthologues on every feature \(i\) of the query genome, normalize these counts by dividing by the largest count, subtract each normalized value from one, sum the results, and divide by the number \(n\) of query features minus one.

$$\frac{1}{n-1}\sum_{i=1}^{n}\left(1 - \frac{x_i}{\max(x)}\right)$$

Here, the index is extended to the whole genome comparisons by computing it for each feature of the target genome and returning the average weighted by feature length.

Note

Note that calculating the tau index on whole-genome nucleotide alignments is not expected to produce meaningful results. This function is more useful when comparing the positions of protein orthologues.

References

Yanai I, Benjamin H, Shmoish M, Chalifa-Caspi V, Shklar M, Ophir R, Bar-Even A, Horn-Saban S, Safran M, Domany E, Lancet D, Shmueli O. (2005). Genome-wide midrange transcription profiles reveal expression level relationships in human tissue specification. Bioinformatics, 21(5):650-659. doi:10.1093/bioinformatics/bti042 . PMID: 15388519

Ranz JM, González PM, Su RN, Bedford SJ, Abreu-Goodger C, Markow T. (2022). Multiscale analysis of the randomization limits of the chromosomal gene organization between Lepidoptera and Diptera. Proc Biol Sci, 289(1967):20212183. doi:10.1098/rspb.2021.2183 . PMID: 35042416

See also

Author

Charles Plessy

Examples

tau_index(exampleTranslocation)
#> [1] 0.5
GenomicBreaks:::.tau_index(c(0,8,0,0,0,2,0,2,0,0,0))
#> [1] 0.95