The GenomicBreaks class

The GBreaks class is a simple wrapper to the GRanges class.

Details

Aligned sequences of the target genome are represented as the main ranges of the GRanges object, and their counterparts in the query genome are represented as a GRanges object sored in the the metadata column query.

Examples

gb <- load_genomic_breaks(system.file("extdata/SacCer3__SacPar.gff3.gz", package = "GenomicBreaks"))
gb

#> GBreaks object with 505 ranges and 2 metadata columns:
#>         seqnames      ranges strand |     score                   query
#>            <Rle>   <IRanges>  <Rle> | <numeric>               <GRanges>
#>     [1]     chrI  5860-10010      + |     10609  NC_047487.1:8723-12716
#>     [2]     chrI 11157-11782      + |      1253 NC_047487.1:13035-13680
#>     [3]     chrI 25371-26528      + |      3651 NC_047487.1:15380-16537
#>     [4]     chrI 26849-29699      + |      6130 NC_047487.1:17263-20175
#>     [5]     chrI 29937-30607      + |       767 NC_047487.1:20203-20866
#>     ...      ...         ...    ... .       ...                     ...
#>   [501]     chrM 67039-67870      + |      1441   NC_018044.1:7758-8439
#>   [502]     chrM 68205-68580      + |       830   NC_018044.1:8783-9180
#>   [503]     chrM 69178-76168      + |     14528  NC_018044.1:9650-16261
#>   [504]     chrM 77356-80022      + |      8066 NC_018044.1:53206-55865
#>   [505]     chrM 80919-85779      + |      6712 NC_018044.1:57885-61592
#>   -------
#>   seqinfo: 17 sequences from an unspecified genome; no seqlengths

gb$query

#> GRanges object with 505 ranges and 0 metadata columns:
#>            seqnames      ranges strand
#>               <Rle>   <IRanges>  <Rle>
#>     [1] NC_047487.1  8723-12716      *
#>     [2] NC_047487.1 13035-13680      *
#>     [3] NC_047487.1 15380-16537      *
#>     [4] NC_047487.1 17263-20175      *
#>     [5] NC_047487.1 20203-20866      *
#>     ...         ...         ...    ...
#>   [501] NC_018044.1   7758-8439      *
#>   [502] NC_018044.1   8783-9180      *
#>   [503] NC_018044.1  9650-16261      *
#>   [504] NC_018044.1 53206-55865      *
#>   [505] NC_018044.1 57885-61592      *
#>   -------
#>   seqinfo: 17 sequences from an unspecified genome; no seqlengths