Breakpoint Pairwise Comparison using Reference Genome as Coordinate System

This function takes two GBreaks objects, with the same target genome. A heatmap is produced, where the breakpoints of query 1 are the centre point reference, and the breakpoints of query 2 that fall within the window are projected onto it.

Usage

bp_pair_analysis(gb1, gb2, window, label)

Arguments

gb1

GBreaks object of the alignment between the target genome and query genome 1.

gb2

GBreaks object of the alignment between the target genome and query genome 2.

window

range around query1 breakpoints of which query2 breakpoints are projected on to. Should be even number

label

string which will be the label on the graph

Value

Returns a heatmaps::Heatmap object of pattern that can be piped into heatmaps::smoothHeatmap and then heatmaps::plotHeatmapList or heatmaps::plotHeatmapMeta.

See also

Author

Charlotte West