Breakpoint graph properties

Computes the breakpoint graph of a pair of chromosomes and returns key properties of the graph, which are used in various rearrangement estimates.

Usage

breakpointGraphProperties(gb)

Arguments

gb

A GBreaks object.

Value

A list containing the following properties of the breakpoint graph: 1. N : the total number of places where a breakpoint could occur (which is the same as the number_aligned_blocks + 1); 2. nbBreakpoints : the total number of breakpoints in the extended permutation; 3. nbCycles : the total number of cycles in the breakpoint graph (trivial and non-trivial cycles).

References

Hannenhalli, Sridhar, and Pavel A. Pevzner. "Transforming cabbage into turnip: polynomial algorithm for sorting signed permutations by reversals." Journal of the ACM (JACM) 46.1 (1999): 1-27.

See also

Author

Priscila Biller