Flag colinear alignments

Flags alignments that are colinear with the next one in sequence order. The flag is added to the first alignment. Colinearity is defined by the fact that the next alignment on the target genome is the same as the next alignment on the query genome, with strand information being taken into account.

flagColinearAlignments(gb, tol = Inf, minwidth = 0, details = FALSE)

Arguments

gb
GBreaks object of the pairwise alignment.

tol
Unaligned regions larger than this tolerance threshold will interrupt colinearity.

minwidth
Remove the intervals whose width smaller than this value.

details
Returns more metadata columns if TRUE.

Value

Returns a modified GBreaks object with a new colinear metadata column indicating if an alignment is colinear with the next one. If the details option is set to TRUE, it will also output the relative position of the following and previous alignments on the target and query genomes (tfoll, tprev, qfoll, qprev), a partial colinearity flag for each genome (t_col and q_col), and the distance to the next alignment on each genome (tdist and qdist).

Details

Internally, flagColinearAlignments() uses the precede and follow functions functions to determine what is the next. For a given range, these functions return the index position of the range it precedes or follows, or NA as the first range follows nothing and the last range precedes nothing. See the examples for details.

Note

The flags are only valid as long as the Genomic Breaks object is not modified by removing alignments or sorting them in a different order.

Pay attention that if the mindwidth option is passed, some intervals are discarded from the returned object. This parameter might be removed in the future if too confusing or useless.

See also

Author

Charlotte West

Charles Plessy

Examples

# Ranges on the plus strand that should coalesce
gb1       <- GRanges(c(A="Ref:100-200:+", B="Ref:400-500:+"))
gb1$query <- GRanges(c(A="Que:100-200",   B="Que:400-500"))
gb1

#> GRanges object with 2 ranges and 1 metadata column:
#>     seqnames    ranges strand |       query
#>        <Rle> <IRanges>  <Rle> |   <GRanges>
#>   A      Ref   100-200      + | Que:100-200
#>   B      Ref   400-500      + | Que:400-500
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

flagColinearAlignments(gb1)

#> GRanges object with 2 ranges and 2 metadata columns:
#>     seqnames    ranges strand |       query colinear
#>        <Rle> <IRanges>  <Rle> |   <GRanges>    <Rle>
#>   A      Ref   100-200      + | Que:100-200     TRUE
#>   B      Ref   400-500      + | Que:400-500    FALSE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Reference range [1] precedes reference range [2]
precede(gb1)

#> [1]  2 NA

# Query range [1] precedes query range [2]
precede(gb1$query)

#> [1]  2 NA

# Ranges on the minus strand that should coalesce
gb2       <- GRanges(c(B="Ref:100-200:-", A="Ref:400-500:-"))
gb2$query <- GRanges(c(B="Que:400-500",   A="Que:100-200"))
gb2$qname <- names(gb2$query)
gb2

#> GRanges object with 2 ranges and 2 metadata columns:
#>     seqnames    ranges strand |       query       qname
#>        <Rle> <IRanges>  <Rle> |   <GRanges> <character>
#>   B      Ref   100-200      - | Que:400-500           B
#>   A      Ref   400-500      - | Que:100-200           A
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Reference range [1] follows reference range [2]
follow(gb2)

#> [1]  2 NA

# Or, ignoring strand, reference range [1] precedes reference range [2]
precede(gb2, ignore.strand = TRUE)

#> [1]  2 NA

# Query range [1] follows query range [2]
follow(gb2$query)

#> [1]  2 NA

flagColinearAlignments(gb2)

#> GRanges object with 2 ranges and 3 metadata columns:
#>     seqnames    ranges strand |       query       qname colinear
#>        <Rle> <IRanges>  <Rle> |   <GRanges> <character>    <Rle>
#>   B      Ref   100-200      - | Que:400-500           B     TRUE
#>   A      Ref   400-500      - | Que:100-200           A    FALSE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Ranges on the minus strand that should not coalesce because they are not
# ordered properly
gb3       <- GRanges(c("Ref:100-200:-", "Ref:400-500:-"))
gb3$query <- GRanges(c("Que:100-200",   "Que:400-500"))
# Reference range [1] follows reference range [2]
follow(gb3)

#> [1]  2 NA

# Query range [1] follows query range [2]
precede(gb3$query)

#> [1]  2 NA

flagColinearAlignments(gb3)

#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |       query colinear
#>          <Rle> <IRanges>  <Rle> |   <GRanges>    <Rle>
#>   [1]      Ref   100-200      - | Que:100-200    FALSE
#>   [2]      Ref   400-500      - | Que:400-500    FALSE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Ranges on the plus strand that should not coalesce because they are not
# ordered properly
gb4       <- GRanges(c("Ref:100-200:+", "Ref:400-500:+", "Ref:600-700:+"))
gb4$query <- GRanges(c("Que:1100-1200:+", "Que:1700-1800:+", "Que:1500-1600:+"))
flagColinearAlignments(gb4)

#> GRanges object with 3 ranges and 2 metadata columns:
#>       seqnames    ranges strand |           query colinear
#>          <Rle> <IRanges>  <Rle> |       <GRanges>    <Rle>
#>   [1]      Ref   100-200      + | Que:1100-1200:+    FALSE
#>   [2]      Ref   400-500      + | Que:1700-1800:+    FALSE
#>   [3]      Ref   600-700      + | Que:1500-1600:+    FALSE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths