Breakpoints

Given a GRanges object, the function produces a GRanges object detailing the breakpoints only. The only converted data is that of the main GRanges subject, and no metadata is processed or carried through

get_bps(
  gr_ob,
  direction = c("both", "left", "right"),
  stranded = FALSE,
  sorted = TRUE
)

Arguments

gr_ob
GRanges object containing pairwise alignment

direction
Will return the breakpoints on both, left or right side(s) of the range.

stranded
If TRUE, will assign a + strand to the left-side breakpoints and a - strand to the right-side ones.

sorted
Sorts the object before returning it.

Value

GRanges object of the breakpoints

Examples

gr <- GRanges(c("chr2", "chr2", "chr1", "chr3"), IRanges::IRanges(1:4, width=4:1))
get_bps(gr)

#> GRanges object with 8 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2         1      *
#>   [2]     chr2         2      *
#>   [3]     chr2         4      *
#>   [4]     chr2         4      *
#>   [5]     chr1         3      *
#>   [6]     chr1         4      *
#>   [7]     chr3         4      *
#>   [8]     chr3         4      *
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths

get_bps(gr, direction = "left")

#> GRanges object with 4 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2         1      *
#>   [2]     chr2         2      *
#>   [3]     chr1         3      *
#>   [4]     chr3         4      *
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths

get_bps(gr, stranded = TRUE)

#> GRanges object with 8 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2         1      +
#>   [2]     chr2         2      +
#>   [3]     chr2         4      -
#>   [4]     chr2         4      -
#>   [5]     chr1         3      +
#>   [6]     chr1         4      -
#>   [7]     chr3         4      +
#>   [8]     chr3         4      -
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths

get_bps(gr, direction = "right", stranded = TRUE)

#> GRanges object with 4 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2         4      -
#>   [2]     chr2         4      -
#>   [3]     chr1         4      -
#>   [4]     chr3         4      -
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths