Load pairwise genome alignments

Loads alignments of a query genome to a target genome from a text file in GFF3 format. By convention, the target genome is the one that was indexed by the aligner.

load_genomic_breaks(
  file,
  target_bsgenome = NULL,
  query_bsgenome = NULL,
  sort = TRUE,
  type = "match_part"
)

Arguments

file

Path to a file in GFF3 format.

target_bsgenome

A BSgenome object representing the target genome.

query_bsgenome

A BSgenome object representing the query genome.

sort

Returns the object sorted, ignoring strand information.

type

Sequence ontology term representing an alignment block (default: match_part).

Value

Returns a GBreaks object where each element represents a pairwise alignment block. The granges part of the object contains the coordinates on the target genome, and the query metadata column contains the querycoordinates in GRanges format. The seqinfo of each BSgenome object passed as parameters are copied to the GRanges objects accordingly.

Details

This function expects the pairwise alignment to be represented in GFF3 format in the following way:

Alignments blocks are represented by entries in specific sequence ontology term in the type column. Other entries will be discarded. The default type is match_part.

The coordinate system of the file is the one of the target genome.

The Target tag in the attribute column contains the coordinates of the match in the query genome. (Sorry that it is confusing…)

Stand information is set so that query genome coordinates are always on the plus strand.

Examples

if (FALSE) {
library("BSgenome.Odioica.local.OSKA2016")
library("BSgenome.Odioica.local.OKI2018.I69")
load_genomic_breaks(
  system.file("extdata/OSKA2016__I69-5.gff3.gz", package="GenomicBreaks"),
  target = OSKA2016,
  query = OKI2018_I69)
}