Find longest matches from target to query genome.

Using a GenomicBreaks object representing the alignment of a query genome on a target genome, finds the longest match of each sequence (representing contigs, scaffolds, etc.) of the query genome on the target genome. Thus, each sequence of the query genome is represented only once in total, but sequences of the target genome can be represented multiple times if they are the longest match of multiple query genome sequences. When the target genome is more contiguous than the query genome, and if there are no major structural variations between them, this will reveal arrangements of colinear sequences in the query genome.

longestMatchesInTarget(gb, min.width = 10000, min.matches = 2)

Arguments

gb
A GenomicBreaks object

min.width
Minimum width of a match (on the query genome) for being considered.

min.matches
Discard query sequences that have fewer longest matches than min.matches on the target. Default is 2, so that only results relevant to chaining genomes are kept.

Value

Returns a GRangesList object containing one GenomicBreak object per sequence on the query genome.

Details

For a more compact version of the results, the output of this function can be piped to strandNames(query = TRUE).

Examples

if (FALSE) {
library("BSgenome.Odioica.local.OSKA2016")
library("BSgenome.Odioica.local.OKI2018.I69")
gb <- load_genomic_breaks(
  system.file("extdata/OSKA2016__I69-5.gff3.gz", package="GenomicBreaks"),
  target = OSKA2016,
  query = OKI2018_I69)
}