Master Breakpoint Analysis

This function brings together the different processes and analyses of the GenomicBreaks package; coalescents, breakpoint coverage and breakpoint proximity to tandem repeats. Input fields can be left blank if the information is not available, and corresponding analyses will not be outputed.

master_bp_analysis(
  pair_gr,
  co_tol,
  ref_tan,
  ref_tan_tol,
  q_tan,
  q_tan_tol,
  ref_cov,
  ref_cov_tol,
  q_cov,
  q_cov_tol
)

Arguments

pair_gr
GRanges object containing pairwise alignment

co_tol
tolerance over which the coalescing algorithm in coalesce_contigs should bridge. If left blank, coalescing will not be performed

ref_tan
tandem repeat coverage for the reference genome in GRanges format

ref_tan_tol
proximity to which a breakpoint should be considered to be "near" a tandem repeat

q_tan
tandem repeat coverage for the query genome in GRanges format

q_tan_tol
proximity to which a breakpoint should be considered to be "near" a tandem repeat

ref_cov
reference genome coverage (per base)

ref_cov_tol
range overwhich to give average coverage over a breakpoint (reference genome)

q_cov
query genome coverage (per base)

q_cov_tol
range overwhich to give average coverage over a breakpoint (query genome)

Value

a list containing two GRanges objects; one for the reference genome and one for the query genome. Metadata columns contain information per breakpoint, indicated by left (start) breakpoint and right (end) breakpoint